+++ datakit/__init__.py
@@ -1,1 +1,4 @@
-__version__='0.1'
+from .command_helpers import CommandHelpers
+
+
+__version__ = '0.1'