+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+Unreleased
+-----------
+
+* Add :py:class:`datakit.command_helpers.CommandHelpers` mixin and misc utility functions in :py:mod:`datakit.utils`.
+
+
 0.1.0 (2017-02-01)
 ------------------